using black to reformat the project
Black is a python code formatter with very few configuration options.
This is useful because it removes the hand reformatting differences
throughout the project and makes decisions about things that tend to
cause disagreements.  Rather than try to come to some sort of consensus,
the tool just makes changes that meet pep8 guidelines for the most part.
Another benefit is that if the entire project is forced into a
formatting standard, changes will only reflect functional things rather
than reformatting things, and thus, diffs will be easier to read.